backend/requirements.txt
numpy
scikit-learn

 groq==0.22.0
 speechrecognition==3.8.1
 
# motor>=3.1.1
 # Testing
 httpx>=0.23.0,<1.0.0
 fastapi==0.104.1
 uvicorn==0.24.0
 python-multipart==0.0.6
 python-dotenv==1.0.0
 pydantic==2.5.0
aiofiles
pyttsx3 
 # Database
# motor==3.1.1
pymongo
 requests==2.28.2
pydub
pyjwt
passlib==1.7.4
tensorflow==2.18.0
Pillow==10.1.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
joblib==1.3.2
pydantic[email]
